--- app/controllers/reservations_controller.rb
@@ -1,23 +1,0 @@
-class ReservationsController < ApplicationController
-  
-  def index
-    @reservations = current_user.reservations
-    render json:  @reservations
-  end
-
-  def create
-    reservation = Reservation.new(reservation_params)
-    reservation.user_id = current_user.id
-    if reservation.save
-      render json: reservation, status: :created
-    else
-      render json: reservation.errors, status: :unprocessable_entity
-    end
-  end
-
-  private
-
-  def reservation_params
-    params.require(:reservation).permit(:city, :date, :room_id)
-  end
-end
