--- app/controllers/rooms_controller.rb
@@ -1,8 +1,0 @@
-class RoomsController < ApplicationController
-
-  skip_before_action :authorize_request, only: :index
-
-  def index
-    render json: Room.all
-  end
-end
