+++ app/controllers/v1/rooms_controller.rb
@@ -1,0 +1,10 @@
+module V1
+  class RoomsController < ApplicationController
+
+    skip_before_action :authorize_request, only: :index
+
+    def index
+      render json: Room.all
+    end
+  end
+end
